+++ minify-html/src/minify/attr.rs
@@ -292,6 +292,8 @@ pub fn minify_attr(
 ) -> AttrMinified {
   let attr_cfg = ATTRS.get(ns, tag, name);
 
+  let do_not_omit = cfg.keep_input_type_text_attr && tag == b"input" && name == b"type" && value_raw.eq_ignore_ascii_case(b"text");
+
   let should_collapse = attr_cfg.filter(|attr| attr.collapse).is_some();
   let should_trim = attr_cfg.filter(|attr| attr.trim).is_some();
   let should_lowercase = attr_cfg.filter(|attr| attr.case_insensitive).is_some();
@@ -341,10 +343,11 @@ pub fn minify_attr(
     value_raw.make_ascii_lowercase();
   };
 
-  if (value_raw.is_empty() && redundant_if_empty)
-    || default_value.filter(|dv| dv == &value_raw).is_some()
-    || (tag == b"script" && name == b"type" && JAVASCRIPT_MIME_TYPES.contains(value_raw.as_slice()))
-  {
+  if !do_not_omit && (
+    (value_raw.is_empty() && redundant_if_empty)
+      || default_value.filter(|dv| dv == &value_raw).is_some()
+      || (tag == b"script" && name == b"type" && JAVASCRIPT_MIME_TYPES.contains(value_raw.as_slice()))
+  ) {
     return AttrMinified::Redundant;
   };