+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # minify-html changelog
 
+## Pending
+
+- Add `keep_input_type_text_attr` option to keep `type=text` on `<input>` elements.
+
 ## 0.14.0
 
 - Add new options to parse and preserve common templating syntax in content source code. NOTE: The parsing is "dumb" and merely looks for the next subsequence in the source code that matches the closing delimiter characters. This means that literal closing delimiter characters (e.g. strings) and nesting may cause parsing to be incorrect.